Update Directory Structure
* Move app files into app
* Move config file into config
* Add sample rack test file
* Add sample config.ru
