+++ Gemfile
@@ -4,3 +4,4 @@ gem 'pry-byebug'
 gem 'faraday'
 gem 'nokogiri'
 gem 'rspec'
+gem 'rack-test'
