+++ app/sync_gis_tables_with_carto.rb
@@ -1,20 +1,42 @@
 #!/usr/bin/env ruby
 require 'active_record'
 require 'faraday'
-require 'pry-byebug'
 require 'yaml'
+require 'pry-byebug'
 
-@settings = YAML.load_file('settings.yml')
+@settings = YAML.load_file('../config/settings.yml')
 
 ActiveRecord::Base.establish_connection(
   adapter:  'postgresql',
   host: @settings['host'],
   port: @settings['port'],
-  database: @settings['tabular_database'],
+  database: @settings['gis_database'],
   username: @settings['database_username'],
   password: @settings['database_password'],
-  schema_search_path: @settings['tabular_schema']
+  schema_search_path: @settings['gis_schema']
 )
+
+def columns_in_table(name, schema=@settings['gis_schema'])
+  sql = <<~SQL
+    SELECT
+      column_name
+    FROM
+      information_schema.columns
+    WHERE
+      table_schema = '#{schema}'
+    AND
+      table_name   = '#{name}';
+  SQL
+  ActiveRecord::Base.connection.execute(sql).pluck('column_name')
+end
+
+def sql_query(table)
+  columns = columns_in_table(table).reject { |column_name| column_name == 'shape' }
+  "SELECT " +
+  columns.join(', ') +
+  ', sde.ST_AsText(sde.ST_Transform(shape, 4326)) AS the_geom' +
+  " FROM #{@settings['gis_schema']}.#{table};"
+end
 
 def record_item_queue(table, id)
   open('import_list.csv', 'a') do |f|
@@ -22,7 +44,7 @@ def record_item_queue(table, id)
   end
 end
 
-def add_carto_sync_for(table, schema=@settings['tabular_schema'])
+def add_carto_sync_for(table, schema=@settings['gis_schema'])
   response = Faraday.post do |req|
     req.url "#{@settings['carto_url']}/api/v1/synchronizations/"
     req.params['api_key'] = @settings['carto_api_key']
@@ -31,12 +53,13 @@ def add_carto_sync_for(table, schema=@settings['tabular_schema'])
                     "provider": "postgres",
                     "connection": {
                       "server": @settings['host'],
-                      "database": @settings['tabular_database'],
+                      "database": @settings['gis_database'],
                       "port": @settings['port'],
                       "username": @settings['database_username'],
                       "password": @settings['database_password']
                     },
                     "table": table,
+                    "sql_query": sql_query(table),
                     "schema": schema
                   },
                   "interval": 2592000
@@ -47,7 +70,8 @@ def add_carto_sync_for(table, schema=@settings['tabular_schema'])
   record_item_queue(table, parsed_body['data_import']['item_queue_id'])
 end
 
-def tables_with_permission(schema=@settings['tabular_schema'])
+# Need to make sure all tables we want have permission for viewer user.
+def tables_with_permission(schema=@settings['gis_schema'])
   sql = <<~SQL
   SELECT
       tablename
@@ -73,4 +97,5 @@ end
 tables_to_sync = ActiveRecord::Base.connection.tables - carto_tables
 no_permission_to_sync = ActiveRecord::Base.connection.tables - tables_with_permission
 tables_with_permission_to_sync = tables_to_sync - no_permission_to_sync
-tables_with_permission_to_sync.first(1).each { |table| add_carto_sync_for(table) }
+binding.pry
+tables_with_permission_to_sync.each { |table| add_carto_sync_for(table) }
