+++ config.ru
@@ -1,0 +1,2 @@
+require_relative 'app/api'
+run MetadataService::API.new
